[#111] refactor : Code separation
