+++ src/components/error/index.ts
@@ -1,1 +1,2 @@
 export { ErrorComponent } from "./ErrorComponent";
+export type { ErrorComponentProps } from "./types";
