+++ src/components/error/types.ts
@@ -1,0 +1,12 @@
+type ErrorType = {
+  message: string;
+  detailMessage?: string;
+  fixText?: string;
+  func?: () => void;
+};
+
+export type ErrorComponentProps = {
+  resetError: () => void;
+  error: ErrorType;
+  errorImg?: string;
+};
